fix: extract is_harness_ready with unit tests to catch && -> || mutation
